Added Rating system, Temporarily remove List Drawer and Watch button, handled button actions when not logged in

+++ app/movie/components/actions.ts
@@ -2,6 +2,7 @@
 
 import { createClient } from "@/utils/supabase/server";
 import { revalidatePath } from "next/cache";
+import { redirect } from "next/navigation";
 
 export async function addOrRemoveFromLibrary(formData: FormData) {
   const user_id = formData.get("user_id");
@@ -12,7 +13,11 @@ export async function addOrRemoveFromLibrary(formData: FormData) {
 
   const supabase = createClient();
 
-  if (isInLibrary === "true") {
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
+
+  if (user && isInLibrary === "true") {
     const { error } = await supabase
       .from("item_lists")
       .delete()
@@ -23,7 +28,7 @@ export async function addOrRemoveFromLibrary(formData: FormData) {
       console.error("Failed to remove from library:", error.message);
     }
     revalidatePath(`/${item_type}/${item_id}`);
-  } else {
+  } else if (user && isInLibrary === "false") {
     const { error } = await supabase
       .from("item_lists")
       .insert([{ user_id, item_id, list_type, item_type }]);
@@ -32,5 +37,39 @@ export async function addOrRemoveFromLibrary(formData: FormData) {
       console.error("Failed to add to library:", error.message);
     }
     revalidatePath(`/${item_type}/${item_id}`);
+  } else {
+    redirect("/login");
+  }
+}
+
+export async function setRating(
+  rating: number,
+  item_id: number,
+  item_type: string
+) {
+  const supabase = createClient();
+
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
+
+  if (user) {
+    const { error } = await supabase.from("item_lists").upsert(
+      {
+        user_id: user.id,
+        item_id,
+        rating,
+        item_type,
+        list_type: "Library",
+      },
+      { onConflict: "user_id, item_id" }
+    );
+
+    if (error) {
+      console.error("Failed to set rating:", error.message);
+    }
+    revalidatePath(`/${item_type}/${item_id}`);
+  } else {
+    redirect("/login");
   }
 }
